Splitting test_commands in different test files.
